完善notes&tag 页面

+++ src/views/index/bloglist.vue
@@ -89,10 +89,26 @@ export default {
       }
     },
     handleDelete(index, row) {
-      console.log(index, row);
-      deletePostById(row.id).then(res => {
-        console.log(res);
-      });
+      this.$confirm("此操作将永久删除该文件, 是否继续?", "提示", {
+        confirmButtonText: "确定",
+        cancelButtonText: "取消",
+        type: "warning"
+      })
+        .then(() => {
+          deletePostById(row.id).then(res => {
+            this.tableData.splice(index, 1);
+            this.$message({
+              message: res.data.result,
+              type: "success"
+            });
+          });
+        })
+        .catch(() => {
+          this.$message({
+            type: "info",
+            message: "已取消删除"
+          });
+        });
     }
   }
 };
